refactor(api): split member-photo.service.ts into services/member/photos/ (api-9si2)
