+++ .beans/api-9si2--refactor-member-photoservicets-589-loc-into-servic.md
@@ -1,10 +1,11 @@
 ---
 # api-9si2
 title: Refactor member-photo.service.ts (589 LOC) into services/member-photo/
-status: todo
+status: in-progress
 type: task
+priority: normal
 created_at: 2026-04-21T13:56:56Z
-updated_at: 2026-04-21T13:56:56Z
+updated_at: 2026-04-21T21:29:34Z
 parent: api-6l1q
 ---
 
@@ -39,3 +40,10 @@ Currently concentrates photo upload / blob linkage / member denormalization in a
 ## Parallelization
 
 No cross-blockers with other service refactor beans — safe to run in a worktree agent concurrently with siblings.
+
+## Findings
+
+- services/member-photo.service.ts:3 — imported `brandId` but was unused post-split (only needed in internal.ts's toPhotoResult); removed during split — low
+- No circular imports; verb files import from ./internal.js only; no back-edges — low
+- Integration test harness in api project produces "DB timeout" log noise for route tests that lack a live PG (428/428 test files pass, 5331/5331 tests pass); pre-existing, unrelated to this refactor — low
+- Max verb file LOC: lifecycle.ts at 204 (well under 300 target); total 627 vs. original 589 due to per-file imports — low
