+++ apps/api/src/__tests__/routes/members/photos/create.test.ts
@@ -13,7 +13,7 @@ import type { ApiErrorResponse } from "@pluralscape/types";
 
 // ── Mocks ────────────────────────────────────────────────────────
 
-vi.mock("../../../../services/member-photo.service.js", () => ({
+vi.mock("../../../../services/member/photos/create.js", () => ({
   createMemberPhoto: vi.fn(),
 }));
 
@@ -29,7 +29,7 @@ vi.mock("../../../../middleware/auth.js", () => mockAuthFactory());
 // ── Imports after mocks ──────────────────────────────────────────
 
 const { createAuditWriter } = await import("../../../../lib/audit-writer.js");
-const { createMemberPhoto } = await import("../../../../services/member-photo.service.js");
+const { createMemberPhoto } = await import("../../../../services/member/photos/create.js");
 const { systemRoutes } = await import("../../../../routes/systems/index.js");
 
 // ── Helpers ──────────────────────────────────────────────────────
