+++ apps/api/src/routes/members/photos/archive.ts
@@ -6,7 +6,7 @@ import { createAuditWriter } from "../../../lib/audit-writer.js";
 import { getDb } from "../../../lib/db.js";
 import { parseIdParam, requireIdParam } from "../../../lib/id-param.js";
 import { createCategoryRateLimiter } from "../../../middleware/rate-limit.js";
-import { archiveMemberPhoto } from "../../../services/member-photo.service.js";
+import { archiveMemberPhoto } from "../../../services/member/photos/lifecycle.js";
 
 import type { AuthEnv } from "../../../lib/auth-context.js";
