+++ apps/api/src/routes/members/photos/reorder.ts
@@ -7,7 +7,7 @@ import { requireIdParam } from "../../../lib/id-param.js";
 import { parseJsonBody } from "../../../lib/parse-json-body.js";
 import { envelope } from "../../../lib/response.js";
 import { createCategoryRateLimiter } from "../../../middleware/rate-limit.js";
-import { reorderMemberPhotos } from "../../../services/member-photo.service.js";
+import { reorderMemberPhotos } from "../../../services/member/photos/update.js";
 
 import type { AuthEnv } from "../../../lib/auth-context.js";
