+++ apps/api/src/routes/members/photos/restore.ts
@@ -6,7 +6,7 @@ import { getDb } from "../../../lib/db.js";
 import { parseIdParam, requireIdParam } from "../../../lib/id-param.js";
 import { envelope } from "../../../lib/response.js";
 import { createCategoryRateLimiter } from "../../../middleware/rate-limit.js";
-import { restoreMemberPhoto } from "../../../services/member-photo.service.js";
+import { restoreMemberPhoto } from "../../../services/member/photos/lifecycle.js";
 
 import type { AuthEnv } from "../../../lib/auth-context.js";
